pyproject.toml
[project]
name = "ai-terminal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.62.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
]
